update `ilius/is`, and fix `tinygo test`

+++ cal_types/ethiopian/ethiopian_test.go
@@ -107,7 +107,7 @@ func TestToJd(t *testing.T) {
 
 func TestGetMonthLen(t *testing.T) {
 	is := is.New(t).MsgSep(", ")
-	testMap := map[[2]int]int{
+	testMap := map[[2]int]uint8{
 		{2015, 1}:  30,
 		{2015, 2}:  30,
 		{2015, 3}:  30,
